pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
APScheduler = "^3.7.0"
bcrypt = "^3.2.0"
Flask = "^2.0.1"
Flask-Limiter = "^1.4"
Flask-Login = "^0.5.0"
Flask-Mail = "^0.9.1"
flask-redis = "^0.4.0"
Flask-SQLAlchemy = "^2.5.1"
gunicorn = "^20.1.0"
poethepoet = "^0.10.0"
prometheus-flask-exporter = "^0.18.2"
pycryptodome = "^3.10.1"
PyJWT = "^2.1.0"
python = "^3.9.6"
python-dotenv = "^0.19.0"
requests = "^2.26.0"
sentry-sdk = "^1.3.1"
SQLAlchemy = "^1.4.23"
ua-parser = "^0.10.0"
webargs = "^8.0.1"
Werkzeug = "^2.0.1"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
flake8 = "^3.9.2"
isort = "^5.9.3"
psycopg2 = "^2.9.1"
pytest = "^6.2.5"
sqlacodegen = "^2.3.0"

[tool.poe.tasks]
start = { shell = "poetry run gunicorn server.wsgi:app --config server/gunicorn.conf.py" }
dev = { shell = "poetry run python3 -m server.wsgi" }
format = { shell = "cd server && poetry run isort . && poetry run black ." }
test = { shell = "poetry run pytest server" }

[tool.black]
line-length = 79

[tool.isort]
include_trailing_comma = true
use_parentheses = true
multi_line_output = 3
lines_between_types = 1

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
